Normalized mouseViewDirections.
Realized that I had foolishly failed to normalize the directions from
the mouse's viewpoint which means that the search for the closest
projector pixel was biased by the magnitude of these vectors as well
as the directions.

Also plotted a bunch of images corresponding to the x, y, and z components
of the vectors calculated during the forward mapping to confirm that this
mapping is working correctly and realized that with the current parameters,
which were measured by me two days ago, this code predicts that the bottom
row of projector pixels hits the dome more than a third of a meter above the
mouse's position in the z-direction.  This cannot be correct so my next task
is to check the frustum geometry and the calculation of the projector's focal
point.