Modified mirror radius.
The mirror radius was increased from 0.2 to 0.215 because the three camera
movie appeared to be over warped a bit.  The horizon was kind of parabolic
rather than flat.  A mirror radius of 0.215 is the best estimate that could be
obtained by simple observation.  A more detailed calibration methodology would
be required to obtain a better estimate.