Modified make_diamond_image.py.
Changed it so that I can specify a list of pixels on which diamonds will be
centered.  The resulting image is projected on the dome and a picture is taken.
The picture of this image projected on the dome can be used to find the real
forward mapping for these center pixels and then I can do a parameter search to
find the parameter valuess that result in calculated directions for these
pixels that match the real directions.