Switched to polar egocentric coordinates...
and added the ability to save initial weights and training data so
interesting results can be reproduced. Created a function to plot values
of the objective function vs the value of each weight. Added the ability
to plot a histogram of final weights after trining the network numerous times.
This shows how consistently the training method finds a good solution.
Created a function to plot the weights and objective function value
as the network is being trained which shows how fast the network
converges to the final weights. Added the ability to train the network using
batches of fireflies instead of being limited to one at a time.